Saving results to external storage and sharing the result .txt (with JSON in it) now possible

+++ AlphaMonitor/src/main/java/com/restracks/android/ble/Chronometer.java
@@ -121,11 +121,11 @@ public class Chronometer extends TextView {
         int hours = (int)(timeElapsed / (3600 * 1000));
         int remaining = (int)(timeElapsed % (3600 * 1000));
 
-        int minutes = (int)(remaining / (60 * 1000));
-        remaining = (int)(remaining % (60 * 1000));
+        int minutes = remaining / (60 * 1000);
+        remaining = remaining % (60 * 1000);
 
-        int seconds = (int)(remaining / 1000);
-        remaining = (int)(remaining % (1000));
+        int seconds = remaining / 1000;
+        remaining = remaining % (1000);
 
         int milliseconds = (((int)timeElapsed % 1000) / 10);
